Toggle paint mode off when user mouseLeaves the grid
It's important to toggle paint mode off when the user's mouse leaves
the grid. This is because if the user is painting - and their mouse leaves
the grid - and they let go of their click - the paint mode is still on and
they can now paint the board without pressing down the mouse. This was a
subtle bug.