Add warning message when user resizes board to be out of bounds
I picked the lower bounds of 1 (width) and 1 (height) because there needs to be
at least 1 cell (e.g 1x1 grid) for Paint by Emoji to be usable.

I picked the upper bounds 45 (width) and 25 (height). I chose this size because
it's a good middle ground between small and large. It allows users to draw
basic art like shapes (e.g. hearts). Another reason for limiting the size of
the board is that the current code is not very performant. When the board
reaches size 100x100 there is a time delay between user paint and emoji
rendering.

As a todo, I would like to rethink the performance implication of the scripting
after the P1 and P2 tasks are complete.