Perf: Update code to handle painting on large canvas like 100x100
There are two performance implications and tradeoffs.

1) Representing an entire canvas at the root (topmost parent) component is
expensive. For the grid needs to be updated (e.g. a user paints an emoji)
we need to update the grid using React's setState. Since we cannot mutate
the grid cell imperatively, we need to clone our grid, update the cell
and pass our cloned grid (the next state) to React's setState function.

This works fine for small grids like 10x8, but performance degrades quickly
as grid size reaches 100x100.

My new approach involves creating single grid cells that manage their own
state. This is more performant than the initial approach because when a
user paints an emoji, we can update the state for the individual cell that
contains that emoji.

The advantage to this approach is that we do not need to clone the entire
grid anytime the user paints an emoji.

This approach however comes with tradeoffs and new performance implications.

2) The new approach introduced another performance implication. If we have
a grid of 500x500 (250,000 grid cells) that means any state changes caused
at the parent components will trigger a re-render to the entire subtree.
An example is when a user selects a new emoji. The state `activeEmoji` is
updated at the parent component. This update triggers a render of the entire
subtree, which consists of 250,000 grid cells.

To deal with this new performance problem, I wrapped each cell in React.memo
and passed it a custom comparison callback. This callback returns `true` for
almost all state changes caused by parent component rendering. By returning
`true` the grid cell components do not re-render based on their prop changes.
By doing this, we can avoid rendering 250,000 cell components when a user
selects an emoji.

Because these components don't render based on outside state changes, we need
to somehow pass the components vital data. For example, should the cell be
paintable? What emoji should it be painted with?

I used a ref (React.createRef) to thread down important pieces of information
such as `pressed`, `activeEmoji`, and `mode`. A ref in React is typically used
to keep track of DOM nodes. React's new useRef API can be used as a mutable
JavaScript object that persists through the React lifecycle.

See: https://reactjs.org/docs/hooks-faq.html#is-there-something-like-instance-variables

Caveats:
Users can now more efficiently paint on very large board sizes. However
large grid sizes still can slow down initial page load.

Given more time, it might be worth looking into limiting the user to seeing
a portion of the grid and virtualizing the rest using a technique called
windowing.

See: https://reactjs.org/docs/optimizing-performance.html#virtualize-long-lists